SU-14 Renamed packages to have the general functionality first and Spring second
