+++ README.md
@@ -20,7 +20,8 @@ Spring Beans that would be useful in any Spring project. This will replace SDJK,
 
 ## 1.0.3
 ##### upcoming
-####
+#### SU-14
+Renamed packages to have the general functionality first and Spring second
 
 ## 1.0.2
 ##### 05-08-2024
