+++ src/main/java/com/sobetech/common/service/spring/date/package-info.java
@@ -1,0 +1,21 @@
+/**
+ * Created by Sobetech Holdings LLC
+ *
+ * Copyright © 2024 Sobetech Holdings LLC, All Rights Reserved
+ *
+ * This software is supplied under the terms of a license agreement or
+ * nondisclosure agreement with Sobetech Holdings LLC, or one of its affiliates,
+ * and may not be used, disseminated, or distributed except in accordance with
+ * the terms of that agreement.
+ *
+ */
+
+/**
+ * Service to work with dates and times
+ *
+ * @author John Murray
+ *
+ * @since May 1, 2024
+ *
+ */
+package com.sobetech.common.service.spring.date;
