+++ src/main/java/com/sobetech/common/service/spring/package-info.java
@@ -1,0 +1,21 @@
+/**
+ * Created by Sobetech Holdings LLC
+ *
+ * Copyright © 2024 Sobetech Holdings LLC, All Rights Reserved
+ *
+ * This software is supplied under the terms of a license agreement or
+ * nondisclosure agreement with Sobetech Holdings LLC, or one of its affiliates,
+ * and may not be used, disseminated, or distributed except in accordance with
+ * the terms of that agreement.
+ *
+ */
+
+/**
+ * Common and generic Spring service that can be used in any application
+ *
+ * @author John Murray
+ *
+ * @since May 1, 2024
+ *
+ */
+package com.sobetech.common.service.spring;
